feat(azure): implement required status checks
In Azure status checks are genre/name so we use the required status checks to match up to the results for the branch status api.
All required checks must exist and must be successful, otherwise we will return yellow for pending/missing/unknown or red for a failure/error on any statuses.